Remove models to avoid circular import
